refactoring of mappers
