+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -3,7 +3,7 @@ package ru.practicum.android.diploma.di
 import org.koin.dsl.module
 import ru.practicum.android.diploma.favorite.data.impl.FavoriteRepositoryImpl
 import ru.practicum.android.diploma.favorite.domain.FavoriteRepository
-import ru.practicum.android.diploma.favorite.domain.models.VacancyConvertor
+import ru.practicum.android.diploma.util.mappers.VacancyEntityMapper
 
 class RepositoryModule {
 
@@ -13,7 +13,7 @@ class RepositoryModule {
             FavoriteRepositoryImpl(get(), get())
         }
 
-        factory { VacancyConvertor() }
+        factory { VacancyEntityMapper() }
 
 
     }
