+++ app/src/main/java/ru/practicum/android/diploma/favorite/data/impl/FavoriteRepositoryImpl.kt
@@ -3,25 +3,25 @@ package ru.practicum.android.diploma.favorite.data.impl
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import ru.practicum.android.diploma.db.AppDatabase
-import ru.practicum.android.diploma.favorite.domain.models.VacancyConvertor
 import ru.practicum.android.diploma.db.entity.FavoriteVacancyEntity
 import ru.practicum.android.diploma.favorite.domain.FavoriteRepository
-import ru.practicum.android.diploma.favorite.domain.models.Vacancy
+import ru.practicum.android.diploma.search.domain.models.Vacancy
+import ru.practicum.android.diploma.util.mappers.VacancyEntityMapper
 
 
 class FavoriteRepositoryImpl(
     private val appDatabase: AppDatabase,
-    private val vacancyConvertor: VacancyConvertor
+    private val vacancyConvertor: VacancyEntityMapper
 ) : FavoriteRepository {
 
     override suspend fun insertVacancyToFavoriteList(vacancy: Vacancy) {
         appDatabase.favoriteVacancyDao()
-            .insertVacancyToFavoriteList(vacancyConvertor.mapVacancyToEntity(vacancy))
+            .insertVacancyToFavoriteList(vacancyConvertor.vacancyVacancyToEntity(vacancy))
     }
 
     override suspend fun deleteVacancyFromFavoriteList(vacancy: Vacancy) {
         appDatabase.favoriteVacancyDao()
-            .deleteVacancyFromFavoriteList(vacancyConvertor.mapVacancyToEntity(vacancy))
+            .deleteVacancyFromFavoriteList(vacancyConvertor.vacancyVacancyToEntity(vacancy))
     }
 
     override fun getAllFavouriteVacancies(): Flow<List<Vacancy>> = flow {
@@ -32,7 +32,7 @@ class FavoriteRepositoryImpl(
     override suspend fun getFavouriteVacancyById(id: Int): Vacancy? {
         val favoriteVacancyEntity = appDatabase.favoriteVacancyDao().getFavouriteVacancyById(id)
         return if (favoriteVacancyEntity != null) {
-            vacancyConvertor.mapEntityToVacancy(favoriteVacancyEntity)
+            vacancyConvertor.vacancyEntityToVacancy(favoriteVacancyEntity)
         } else {
             //TODO обсудить с Таней, что возвращать на экран деталей если ошибка получения информации из БД
             null
@@ -44,7 +44,7 @@ class FavoriteRepositoryImpl(
     }
 
     private fun convertFromVacancyEntityToVacancy(listOfEntities: List<FavoriteVacancyEntity>): List<Vacancy> {
-        return listOfEntities.map { vacancy -> vacancyConvertor.mapEntityToVacancy(vacancy) }
+        return listOfEntities.map { vacancy -> vacancyConvertor.vacancyEntityToVacancy(vacancy) }
     }
 
 }
