--- app/src/main/java/ru/practicum/android/diploma/favorite/domain/models/Vacancy.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.favorite.domain.models
-
-data class Vacancy(
-val id: Int,
-val name: String,
-//TODO: implement model
-)
