+++ app/src/main/java/ru/practicum/android/diploma/favorite/recycle_view/VacancyAdapter.kt
@@ -1,0 +1,4 @@
+package ru.practicum.android.diploma.favorite.recycle_view
+
+class VacancyAdapter {
+}
