+++ app/src/main/java/ru/practicum/android/diploma/util/mappers/ConverterListOfStringsDB.kt
@@ -1,0 +1,19 @@
+package ru.practicum.android.diploma.util.mappers
+
+import androidx.room.TypeConverter
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+
+class ConverterListOfStringsDB {
+    @TypeConverter
+    fun fromString(value: String): List<String> {
+        val listType = object : TypeToken<List<String>>() {}.type
+        return Gson().fromJson(value, listType)
+    }
+
+    @TypeConverter
+    fun fromArrayList(list: List<String>): String {
+        val gson = Gson()
+        return gson.toJson(list)
+    }
+}
